Added a command to display a list of districts and a command with a tooltip. Bugs in other commands have been fixed.

+++ src/main/java/spring/telegrambot/weatherBot/data/weather/parts/Wind.java
@@ -1,5 +1,5 @@
 package spring.telegrambot.weatherBot.data.weather.parts;
 
-public record Wind (int speed,
-         int deg){
+public record Wind (double speed,
+         double deg){
 }
